requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
black==24.4.2
click==8.1.7
coverage==7.5.3
Django==4.2.13
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
exceptiongroup==1.2.1
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.1
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
uritemplate==4.1.1
zipp==3.19.2
